[data] mock coffee seed data based on my setup

+++ Backend/Entities/Burr.cs
@@ -3,7 +3,7 @@ namespace Backend.Entities;
 public class Burr : BaseEntity
 {
   public string Type { get; set; } = string.Empty;
-  public int Diameter { get; set; }
+  public decimal Diameter { get; set; }
 
   public ICollection<GrinderBurr> CompatibleGrinders { get; set; } = [];
 }
